server/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
mysqlclient>=2.0.3,<3.0
ipdb
djangorestframework-simplejwt[crypto]
isort
